pyproject.toml
[tool.poetry]
name = "distime"
version = "0.1.0"
description = "this bot converts input time to timestamp structure in for discord"
authors = ["Daylight <woodsfiery@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
